+++ ninja-core/src/main/java/ninja/utils/ClockImpl.java
@@ -1,0 +1,10 @@
+package ninja.utils;
+
+public class ClockImpl implements Clock {
+
+    @Override
+    public long currentTimeMillis() {
+        return System.currentTimeMillis();
+    }
+
+}